moved my converter logic to sep file and add an optional command arg --outlined, create a basic test script but did not revert the base_converter file to original because there is some code committed into it but instead renamed some function name which might conflict due to given function name convert_image_to_ascii
